requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

# Basic packages
numpy
pandas
scipy
matplotlib
Pillow
scikit-learn
nlpaug
spacy

################################################################
# TensorFlow v.2.1 and above include both CPU and GPU versions.

tensorflow==2.3.0
keras-self-attention
keras-tuner